fix: further testinf and refinement
