+++ types/turbodocx__html-to-docx/.npmignore
@@ -1,0 +1,5 @@
+*
+!**/*.d.ts
+!**/*.d.cts
+!**/*.d.mts
+!**/*.d.*.ts
